requirements.txt
langchain>=0.2.12
langchain-openai>=0.1.20
langgraph>=0.2.35
python-dotenv>=1.0.1
pydantic>=2.7.4
psycopg[binary]>=3.2.1